Add cache to the image widget (Patch #18 by andrew-101)
"This patch adds a global _cache attribute to the Image widget and a BGUI_CACHE option.
If the user knows a particular image will be loaded many times they can use the BGUI_CACHE option to add it to the cache. I'm not sure how big an ImageFFmpeg object is, I just thought that it might not always be practical to cache every image loaded - hence adding the option.

From use within my game I have noticed a significant drop in loading times.

I'm not sure if the same concept can be applied to a video, so I decided not to include it in the patch. Maybe if there is a way to clone a VideoFFmpeg object it could work."