Adding outline support to Label. There are three options:
OutlineColor = The color to use for the outline
OutlineSize = The size (in pixels) of the outline (0 for no outline)
OutlineSmoothing = Whether or not to attempt to smooth the outline. This creates nicer outlines at the cost of performance. Without smoothing you do 10 draws to render a label with outlines. With smoothing it's OutlineSize^2 + 1 draws.